+++ Domain/Entities/Common/BaseEntity.cs
@@ -38,12 +38,12 @@ public abstract class BaseEntity<TKey> : IBaseEntity<TKey>
     public virtual TKey Id { get; set; }
     public DateTimeOffset CreateDate { get; set; } = DateTimeOffset.Now;
     public DateTimeOffset UpdateDate { get; set; } = DateTimeOffset.Now;
-    public int CreatorUserId { get; set; }
+    public long CreatorUserId { get; set; }
     [IgnoreDataMember]
     public User CreatorUser { get; set; }
 }
 
-public abstract class BaseEntity : BaseEntity<int>
+public abstract class BaseEntity : BaseEntity<long>
 {
     
 }
