+++ Service/Message/IMessageService.cs
@@ -2,6 +2,6 @@
 {
     public interface IMessageService
     {
-        Task<bool> SendMessageAsync(string phoneNumber, string message, int? userId, int? creatorUserId, CancellationToken ct);
+        Task<bool> SendMessageAsync(string phoneNumber, string message, long? userId, long? creatorUserId, CancellationToken ct);
     }
 }
