+++ Service/Model/Contracts/IPasswordHistoryService.cs
@@ -9,7 +9,7 @@ namespace Service.Model.Contracts
 {
     public interface IPasswordHistoryService
     {
-        Task<PasswordHistory> CheckPasswordHistoryAsync(string password, User user, int createrUserId, CancellationToken ct);
+        Task<PasswordHistory> CheckPasswordHistoryAsync(string password, User user, long createrUserId, CancellationToken ct);
         Task AddAsync(PasswordHistory model, CancellationToken ct);
     }
 }
