+++ WebClient/Components/Pages/SessionsUser/SessionsUserIndex.razor.cs
@@ -61,7 +61,7 @@ public partial class SessionsUserIndex : ComponentBase
         await GetData();
     }
 
-    private async Task Disable(int id)
+    private async Task Disable(long id)
     {
         try
         {
